Setup project structure (#1)
Create packages for calling iTwin APIs and common React components.
Setup a minimal iTwin.js test app for developing and testing these
packages.

The test app will attempt and fail to open a snapshot iModel. This is
intended, as the goal was to test whether application frontend can
communicate with backend.

No end-to-end test setup yet, the decision on how to set them up is
deferred until later.