Improve authorization robustness and performance (#17)
I found two problems with existing authorization implementation:

* Vite queues a huge amount of requests for `.ts` files, which forces
authorization to stall until these requests are cleared. Sometimes the
wait is long enough to cause oidc-client-ts to timeout and fail silent
sign-in check.
* Normally we perform a silent sign-in on app startup to authorize the
user in the background. When it fails, we direct the user to sign-in
normally and handle the result on authorization callback URL. The
problem is, after signing-in the user this way, we still are performing
the silent sign-in when the app starts initializing again.

This PR makes the following changes:

* Delay iTwinJsApp loading until the user is signed in. This will reduce
possibility for authorization request to timeout.
* Do not perform a silent sign-in if user has already signed-in in other
way.