Fix "module not found" errors when importing changed-elements-react (#28)
Developers who import @itwin/changed-elements-react package would
encounter "module not found" error in TypeScript language server if we
published this package as it was.

To fix any module resolution errors that might occur, I decided to ship
changed-elements-react package with support for CommonJS and ESM module
types. In ESM, relative imports must specify the file extension (this is
also supported in CommonJS), but this change made Vite and TS server not
function properly in our test app. Both were expecting output `.js`
directories to exist, but we are only transpiling code in memory.

The workarounds that I came up with were:
* Have `types` field in `changed-elements-react/package.json` point to
actually existing development files on the file system. Before
publishing the package, a script updates this field to point to the
transpiled output.
* Rewrite changed-elements-react `.js` imports as equivalent `.ts` ones
as Vite loads these files in memory.

Lastly, I made sure that projects that are still not ready to upgrade to
`"moduleResolution": "node16"` could also import changed-elements-react
without encountering any problems by adding legacy `main` field to
`package.json` which points to our CommonJS output.