Remove dependency on @itwin/appui-react (#51)
We no longer provide a version comparison frontstage that users can
simply add to their application. This was done in order to reduce the
scope of this package and make developing, maintaining, and using this
package easier. We will consider creating a separate package for hosting
a reusable frontstage, should such a need arrive.

The existing frontstage code is moved to the test-app-frontend
temporarily, and will eventually be deleted entirely.