Remove changeset cache (#60)
Version select dialog used to keep all named version and changeset data
that is received the first time the dialog is opened. This could lead to
data inconsistency when this data is updated on the server, and could
not be resolved with the current API that we provide. To solve this
problem, we instead opt to not cache named version and changeset
responses so that version compare dialog always shows fresh data. We
will no rely on consumers to provide a caching implementation if they
have such requirement.

Removed `@itwin/imodels-client-management` dependency because using its
`IModelsClient` in our interface definitions was problematic. It
required users to install the exact package version that we depended on
and caused unnecessary headache. We could have fixed this by moving it
to `peerDependencies`, but having fewer dependencies is an even better
option.