Fixed infinite load that would occur with changed element entry look up (#115)
We should have been doing a cache look-up. We passed an unloaded change
elements entry, which was cached as loaded. So we would see the unloaded
entry and return unloaded and try to load it. Then load it cache we
loaded it. Then check if it was loaded and then get back unloaded bc we
passed in an unloaded entry to check ...

Fixed by using cached value instead