Updated release workflow and reverted changes from last failed publish (#151)
Added debugging output after release action and adjusted flags for
release action.

Fixes the publishing issue. Trigger must be on pushes to master only. 



When a Pr gets merged into the master, that causes a version update,
i.e., has a changeset.

A workflow will pick up and create a release branch if not already
created.

The release branch will version the package and update the change log.
Once the release branch merges into the master again, a package update
will be detected, a tag and release will be created, and the npm package
will be published.



If the release branch is present and active and new changes are merged
into the master from the dev branches, the release branch will be
updated, and no package release will be triggered until the release
branch is merged.



The error causing the failure was due to a manual trigger. This is
supposed to work off pushes to master only. So, I removed the manual
trigger and added debugging logs.