Release packages [publish docs] (#154)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to master, this PR
will be updated.


# Releases
## @itwin/changed-elements-react@0.12.0

### Minor Changes

####
[0.12.0](https://github.com/iTwin/changed-elements-react/tree/v0.12.0/packages/changed-elements-react)
- 2025-01-21

- Added experimental React component for the new Named Version selector.
Its name or API is not stable yet, but you can try it out the following
way.

    ```TypeScript
import { NamedVersionSelectorWidget } from
"@itwin/changed-elements-react/experimental";

    [...]

      return (
        <VersionCompareContext iModelsClient={iModelsClient}>
          <NamedVersionSelectorWidget iModel={iModel} />
        </VersionCompareContext>
      );
    }
    ```

Co-authored-by: imodeljs-admin <imodeljs-admin@users.noreply.github.com>