Release packages [publish docs] (#166)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to master, this PR
will be updated.

# Releases
## @itwin/changed-elements-react@0.12.2

### Patch Changes

####
[0.12.2](https://github.com/iTwin/changed-elements-react/tree/v0.12.2/packages/changed-elements-react)
- 2025-01-28

Added a feedback button to the experimental UI and fixed the spacing on
named versions in the new widget list.
Import moved for experimental UI. As previous way of importing did not
work.
Marked new component as a alpha and and subject to change.

```TypeScript
import { NamedVersionSelectorWidget } from "@itwin/changed-elements-react";

[...]

  return (
    <VersionCompareContext iModelsClient={iModelsClient}>
      <NamedVersionSelectorWidget iModel={iModel} />
    </VersionCompareContext>
  );
}
```

Co-authored-by: imodeljs-admin <imodeljs-admin@users.noreply.github.com>