Release packages [publish docs] (#175)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to master, this PR
will be updated.

# Releases
## @itwin/changed-elements-react@0.14.2

### Patch Changes

####
[0.14.2](https://github.com/iTwin/changed-elements-react/tree/v0.14.2/packages/changed-elements-react)
- 2025-02-14

-   Updates the new experimental widget's font weight and font size.
- Updates the current and target comparison header info to now be shown
on the comparison tree widget.
- Fixes the filter box to fit the content, getting rid of the horizontal
scroll bar.

Co-authored-by: imodeljs-admin <imodeljs-admin@users.noreply.github.com>