Release packages [publish docs] (#177)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to master, this PR
will be updated.

# Releases
## @itwin/changed-elements-react@0.14.3

### Patch Changes

####
[0.14.3](https://github.com/iTwin/changed-elements-react/tree/v0.14.3/packages/changed-elements-react)
- 2025-02-19

UI tweaks for:

-   Font size, weight, and style
-   Property comparison table coloring changes per row
-   Updated colors for elements in tree (squares)

Fixes:

- Forever loading in the elements tree when backing out of property
comparison (experimental widget only bug)

Co-authored-by: imodeljs-admin <imodeljs-admin@users.noreply.github.com>