Release packages [publish docs] (#203)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to master, this PR
will be updated.

# Releases
## @itwin/changed-elements-react@0.14.12

### Patch Changes

####
[0.14.12](https://github.com/iTwin/changed-elements-react/tree/v0.14.12/packages/changed-elements-react)
- 2025-04-09

- Added responsive design for experimental widget. When widget is
smaller size we display a minimal design.

Co-authored-by: imodeljs-admin <imodeljs-admin@users.noreply.github.com>