Release packages [publish docs] (#241)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to master, this PR
will be updated.

# Releases
## @itwin/changed-elements-react@0.15.7

### Patch Changes

####
[0.15.7](https://github.com/iTwin/changed-elements-react/tree/v0.15.7/packages/changed-elements-react)
- 2025-07-11

Changed elements gets hidden when toggling filters

Co-authored-by: imodeljs-admin <imodeljs-admin@users.noreply.github.com>