Release packages [publish docs] (#254)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to master, this PR
will be updated.

# Releases
## @itwin/changed-elements-react@0.17.0

### Minor Changes

####
[0.17.0](https://github.com/iTwin/changed-elements-react/tree/v0.17.0/packages/changed-elements-react)
- 2025-08-12

_Frontend Enhancements:_

1. Change named version list to be a infinite list, so we do not spam
the UI with quires for constructing Named Versions

_Frontend Bug Fixes:_

1. Fixed active version header not clearing once a comparison is
complete
2. Disabled process result buttons when a comparison is currently being
run

Co-authored-by: imodeljs-admin <imodeljs-admin@users.noreply.github.com>