Release packages [publish docs] (#257)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to master, this PR
will be updated.

# Releases
## @itwin/changed-elements-react@0.17.1

### Patch Changes

####
[0.17.1](https://github.com/iTwin/changed-elements-react/tree/v0.17.1/packages/changed-elements-react)
- 2025-09-03

Fixed miss-alignment issue in named versions entries

Co-authored-by: imodeljs-admin <imodeljs-admin@users.noreply.github.com>