Release packages [publish docs] (#265)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to master, this PR
will be updated.

# Releases
## @itwin/changed-elements-react@1.0.0

### Major Changes

####
[1.0.0](https://github.com/iTwin/changed-elements-react/tree/v1.0.0/packages/changed-elements-react)
- 2025-09-16

# Major Change

Updated all Itwin JS to 5.x.x and stopped using deprecated methods.
This is a breaking change and will require updates to consumers of this
package in their implementation of their property comparison frontstage.

Co-authored-by: imodeljs-admin <imodeljs-admin@users.noreply.github.com>